requirements.txt
spacy
flask
uvicorn
pdf2image
pytesseract
python-docx
